Extract `ExerciseWithSetsInfoView` and `ExerciseWithSetInfoDateFocusedView` into their own file for reusable components
